requirements.txt
nose>=1.3.7
PyHamcrest>=1.9.0
schedule>=0.3.2
mockito>=0.7.0
Flask>=0.12
Pillow